Implemented Gates for authorization based on permissions

+++ app/Http/Controllers/RoleController.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Resources\RoleResource;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
@@ -13,6 +14,8 @@ class RoleController extends Controller
     
     public function index()
     {
+        Gate::authorize('view', 'roles');
+
         return RoleResource::collection(Role::all());
     }
 
@@ -20,6 +23,8 @@ class RoleController extends Controller
     
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::create($request->only('name'));
 
         if($permissions = $request->input('permissions')) {
@@ -38,6 +43,8 @@ class RoleController extends Controller
     
     public function show($id)
     {
+        Gate::authorize('view', 'roles');
+
         return new RoleResource(Role::find($id));
     }
 
@@ -45,6 +52,8 @@ class RoleController extends Controller
     
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::find($id);
 
         $role->update($request->only('name'));
@@ -67,9 +76,11 @@ class RoleController extends Controller
     
     public function destroy($id)
     {
-        Role::destroy($id);
+        Gate::authorize('edit', 'roles');
         
         DB::table('role_permission')->where('role_id', $id)->delete();
+        
+        Role::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
     }
